\name{accumulate}
\alias{accumulate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Count photons falling onto a pixel grid to form a matrix image
}
\description{
  Discretises the positions of light photons by binning into nearest CCD wells, to form photoelectron counts.
  Currently it is assumed that each photon always produces exactly 1 electron.
}
\usage{
accumulate(input = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input}{
	A two-column matrix giving the x and y positions in mm of light photons at the CCD face.
}
}
\details{
This function would not normally be called directly by the user, but implicitly through a call to \code{expose} or
similar. Typically the \code{input} light photon positions will have been generated by \code{phosphor} and transformed
to the CCD chip by \code{taper}.
}
\value{
	A matrix of values corresponding to accumulated photoelectrons at each CCD pixel position.
}
\references{
	\url{http://dx.doi.org/10.1107/S0021889810033418}
}
\author{
	David G. Waterman.
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
	\code{\link{detector}},
	\code{\link{expose}}.
}
\examples{
data(detector)
x_pos <- spot()
light_pos <- taper(phosphor(x_pos))
e_image <- accumulate(light_pos)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datagen}

