\name{DISP-internal}
\title{
Internal DISP objects
}
\description{
	Internal DISP objects. Currently \code{.undistortPoints},
	\code{.undistortPixel},
	\code{.makeSpotMask},
	\code{.findCentroids},
	\code{.distortPoints},
	\code{.background}.
}
\details{These are not to be called by the user.}
\keyword{internal}