\name{DISP-package}
\alias{DISP-package}
\alias{DISP}
\docType{package}
\title{
Diffraction Image Statistics Package
DISP
}
\description{
	This package consists of functions that simulate the operation of a phosphor-taper-CCD
	detector of the type commonly used in macromolecular crystallography. In addition, functions
	are provided to perform spot integration on images, for graphical display of images and for
	reading image and reflection files from real-world detectors.
}
\details{
\tabular{ll}{
Package: \tab DISP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-10\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
	David G. Waterman <dgwaterman@gmail.com>

}
\references{
	\url{http://dx.doi.org/10.1107/S0021889810033418}
}

\keyword{ package }

\examples{
#image generation and spot integration is demonstrated in the documentation for
#the dataset spotSimulations
example(spotSimulations)
}
