\name{memory}
\alias{mem}
\alias{size}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Memory usage of objects in the Global Environment
}
\description{
Data used or generated by \pkg{DISP} may take up a large amount of memory. These functions are provided for convenience
allowing simple queries of the sizes of objects in the Global Environment.
}
\usage{
mem()
size()
}
%- maybe also 'usage' for other objects documented here.

\value{
For \code{mem}, no value is returned, but a dotchart showing all objects using > 1MB of storage space in the Global Environment is plotted.
For \code{size}, a named vector of storage sizes in MB for all objects in the Global Environment, in order of decreasing size, is returned and also printed.
}
\references{
\url{http://dx.doi.org/10.1107/S0021889810033418}
}
\author{
David G. Waterman.
}

\examples{
#make some large objects
a <- rnorm(1e6)
b <- cbind(a, a)
size()
mem()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}

