\name{phosphor}
\alias{phosphor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulates light generation from a phosphor
}
\description{
	Given the positions of incident X-ray photons, generates showers of light photons
	distributed around these positions.
}
\usage{
phosphor(input, PSF = detector$PSF, PHS = detector$PHS)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input}{
	A two-column matrix giving the x and y positions in mm of X-ray photons incident
	on the detector face.
}
  \item{PSF}{
	A two-column matrix describing the radial probability for generated light photon positions.
}
  \item{PHS}{
	A two-column matrix giving the scintillation spectrum (\sQuote{Pulse Height Spectrum})
	describing the intensity distribution of light output.
}
}
\details{
	This function relies on parameters in the list \code{detector} being defined.
	
	For normal use of the simulation this function need not be called directly, but rather from
	higher level functions such as \code{expose}.
}
\value{
	A two-column matrix giving the x and y positions in mm of light photons generated by the phosphor
	at the face of the fibre optic taper.
}
\references{
	\url{http://dx.doi.org/10.1107/S0021889810033418}
}
\author{
	David G. Waterman.
}

\seealso{
	\code{\link{detector}},
	\code{\link{expose}},
	\code{\link{exposeSpots}},
	\code{\link{exposeFlood}}.
}
\examples{
data(detector)
x <- phosphor(spot(10))
}

\keyword{datagen}


